refactor: change class name and path
