+++ src/main/java/org/cotato/csquiz/domain/education/scheduler/QuizScheduler.java
@@ -4,7 +4,7 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.cotato.csquiz.domain.education.entity.Quiz;
 import org.cotato.csquiz.domain.education.service.ChoiceService;
-import org.cotato.csquiz.domain.education.service.DiscordService;
+import org.cotato.csquiz.domain.education.service.component.DiscordConnector;
 import org.cotato.csquiz.domain.education.service.component.QuizReader;
 import org.cotato.csquiz.domain.education.service.dto.RandomQuizResponse;
 import org.cotato.csquiz.domain.education.util.DiscordUtil;
@@ -17,7 +17,7 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class QuizScheduler {
 
-    private final DiscordService discordService;
+    private final DiscordConnector discordConnector;
     private final QuizReader quizReader;
     private final ChoiceService choiceService;
 
@@ -27,7 +27,7 @@ public class QuizScheduler {
         Quiz quiz = quizReader.getRandomDiscordQuiz();
         RandomQuizResponse randomQuiz = RandomQuizResponse.from(quiz);
 
-        discordService.sendMessageToTextChannel(DiscordUtil.getMultipleQuizEmbeds(randomQuiz),
+        discordConnector.sendMessageToTextChannel(DiscordUtil.getMultipleQuizEmbeds(randomQuiz),
                 DiscordUtil.getButtons(choiceService.findAllChoices(randomQuiz.id())));
         log.info("[디스코드 채널 문제 전송 완료: {}]", randomQuiz.id());
     }
